requirements.txt
coverage
flask
marrow.mailer
mock
nose
pea
sphinx
